Delete Part Done

+++ src/page/Dashboard/ManageProducts.js
@@ -1,3 +1,6 @@
+import { confirmAlert } from 'react-confirm-alert'; // Import
+import 'react-confirm-alert/src/react-confirm-alert.css'; // Import css
+
 import React from 'react';
 import { useQuery } from 'react-query';
 import { useParams } from 'react-router-dom';
@@ -5,8 +8,6 @@ import { toast } from 'react-toastify';
 import Loading from '../Shared/Loading';
 
 const ManageProducts = () => {
-
-    let { id } = useParams();
 
     let { data: parts, isLoading, refetch } = useQuery('parts', () => fetch('http://localhost:5000/parts', {
         method: 'GET',
@@ -20,24 +21,61 @@ const ManageProducts = () => {
         return <Loading> </Loading>
     }
 
-    let deleteParts = event => {
-        console.log(event);
-        if (event) {
-            <div></div>
-        }
-        fetch(`http://localhost:5000/part/${event}`, {
-            method: 'DELETE',
-            headers: {
-                'content-type': 'application/json'
-            }
-        })
-            .then(res => res.json())
-            .then(data => {
-                console.log(data);
-                toast.warning(' Parts Delete Successfully ')
-                refetch();
-            })
+
+
+    const submit = (id) => {
+
+        confirmAlert({
+            title: 'Confirm to submit',
+            message: 'Are you sure to do this ?',
+            buttons: [
+                {
+                    label: 'Yes',
+                    onClick: () => {
+                        console.log(id);
+                        fetch(`http://localhost:5000/part/${id}`, {
+                            method: 'DELETE',
+                            headers: {
+                                'content-type': 'application/json'
+                            }
+                        })
+                            .then(res => res.json())
+                            .then(data => {
+                                console.log(data);
+                                toast.warning(' Parts Delete Successfully ')
+                                refetch();
+                            })
+                    }
+                },
+                {
+                    label: 'No',
+                    onClick: () =>{
+                        toast.warn(' Delete Cancel ');
+                    }
+                }
+            ]
+        });
     }
+
+
+    // let deleteParts = event => {
+    //     console.log(event);
+    //     // if (event) {
+    //     //     <div></div>
+    //     // }
+    //     fetch(`http://localhost:5000/part/${event}`, {
+    //         method: 'DELETE',
+    //         headers: {
+    //             'content-type': 'application/json'
+    //         }
+    //     })
+    //         .then(res => res.json())
+    //         .then(data => {
+    //             console.log(data);
+    //             toast.warning(' Parts Delete Successfully ')
+    //             refetch();
+    //         })
+    // }
 
     return (
         <div>
@@ -80,10 +118,18 @@ const ManageProducts = () => {
                                             <td> {part.stock} </td>
 
                                             <td> {part.minOrder} </td>
-                                            {/* 
-                                            <td> <button onClick={()=> deleteParts(part._id)} className=' btn bg-red-600'> X </button> </td> */}
 
-                                            <td> <label for="my-modal-6" class="btn modal-button"> X </label>
+                                            {/* <td> <button onClick={() => deleteParts(part._id)} className=' btn bg-red-600'> X </button> </td> */}
+
+                                            <td>
+
+                                                <div className='container'>
+                                                    <button className='btn bg-red-700' onClick={() => submit(part._id)}>Delete </button>
+                                                </div>
+
+                                            </td>
+
+                                            {/* <td> <label for="my-modal-6" class="btn modal-button"> X </label>
                                                 
                                                     <input type="checkbox" id="my-modal-6" class="modal-toggle" />
                                                     <div class="modal modal-bottom sm:modal-middle">
@@ -92,12 +138,11 @@ const ManageProducts = () => {
                                                             <h3 class="font-bold text-xl text-red-600">Are you sure you wants to DELETE this parts</h3>
                                                             <p class="py-4">Once you delete you can never recover the parts. Its Gone Permanently !! </p>
                                                             <div class="modal-action">
-                                                                <label for="my-modal-6" onClick={() => deleteParts(part._id)} class="btn"> Delete </label>
+                                                                <label for="my-modal-6" onClick={(() => deleteParts(part._id))} class="btn"> Delete </label>
                                                             </div>
                                                         </div>
                                                     </div>
-                                               
-                                            </td>
+                                            </td> */}
                                         </tr>
                                     </tbody>)
                                 }
